+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+* Fix widgets failing to update on devices that allow less bitmap memory than
+  expected
+
 ## [v1.46.0] - 2026-09-02
 [v1.46.0]: https://github.com/fibelatti/photo-widget/releases/tag/v1.46.0